\name{simplebgwprocess}
\alias{simplebgwprocess}
\title{BGW Branching process}
\usage{
simplebgwprocess(n, p.vec, n.sim)
}
\description{
Prints 'To generate samples from a simple (BGW) Branching process!'.
}
\examples{
## x.sample is the generated samples of this (BGW) Branching process
## x.sample = bokai.bgw.process(n=10, p.vec=c(1/2, 1/4, 1/4, 0), n.sim=100)
}
